+++ reusable_workflows/tests/test_external_changes.py
@@ -1,7 +1,7 @@
 import pytest
 import subprocess
 from unittest.mock import patch, mock_open
-from security_checks.security_checks import get_changed_files, load_config, check_files_against_blacklist
+from repo_policies.check_external_changes import get_changed_files, load_config, check_files_against_blacklist
 
 
 @patch("subprocess.run")
